requirements.txt
# Core ML
torch
transformers
accelerate
datasets==4.6.1
peft

# Inference
vllm>=0.18.0

# Config management
hydra-core
omegaconf

# Utilities
pyyaml
numpy
pandas
tqdm

matplotlib
scipy
factor_analyzer
seaborn
umap-learn
sentence-transformers
